facotring out close modal hook
