+++ app/javascript/components/search/controls/MoreFacetsButton.js
@@ -4,6 +4,7 @@ import { faSlidersH } from '@fortawesome/free-solid-svg-icons'
 
 import FacetsAccordion from './FacetsAccordion'
 import { StudySearchContext } from 'providers/StudySearchProvider'
+import useCloseableModal from 'hooks/closeableModal'
 
 /**
  * Component for "More Facets" button.  Clicking shows facets accordion box.
@@ -14,32 +15,7 @@ export default function MoreFacetsButton(props) {
   const searchContext = useContext(StudySearchContext)
   const [show, setShow] = useState(false)
 
-  // const facetName = props.facet.name;
-
-  function handleClick() {
-    setShow(!show)
-  }
-
-  // add event listener to detect mouseclicks outside the accordion, so we
-  // know to close it if we have any more controls like this, consider a HOC
-  // or custom hook for this behavior (shared in FacetControl as well)
-  useEffect(() => {
-    // add when mounted
-    document.addEventListener('mousedown', handleOtherClick)
-    // return function to be called when unmounted
-    return () => {
-      document.removeEventListener('mousedown', handleOtherClick)
-    }
-  }, [])
-
-  const node = useRef()
-  const handleOtherClick = e => {
-    if (node.current.contains(e.target)) {
-      // click was inside the modal, do nothing
-      return
-    }
-    setShow(false)
-  }
+  const { node, clearNode, handleButtonClick } = useCloseableModal(show, setShow)
 
   const numFacetsApplied = props.facets.filter(facet => {
     const facets = searchContext.params.facets
@@ -53,7 +29,7 @@ export default function MoreFacetsButton(props) {
       className={`${show || numFacetsApplied ? 'active' : ''} facet`}
       ref={node}>
       <a
-        onClick={handleClick}>
+        onClick={handleButtonClick}>
         <FontAwesomeIcon className="icon-left" icon={faSlidersH}/>
           More Facets { facetCountString }
       </a>
